update class package: move to interceptor package
