+++ app/src/main/java/com/myduka/app/api/interceptor/AccessTokenIterceptor.java
@@ -1,8 +1,11 @@
-package com.myduka.app.api;
-
+package com.myduka.app.api.interceptor;
 
 
 import android.support.annotation.NonNull;
+import android.util.Base64;
+
+import com.myduka.app.BuildConfig;
+import com.myduka.app.api.ApiClient;
 
 import java.io.IOException;
 
@@ -16,18 +19,19 @@ import okhttp3.Response;
  *
  * @author Thomas Kioko
  */
-public class AuthInterceptor implements Interceptor {
+public class AccessTokenIterceptor implements Interceptor {
 
-    private String mAuthToken;
+    AccessTokenIterceptor() {
 
-    public AuthInterceptor(String authToken) {
-        mAuthToken = authToken;
     }
 
     @Override
     public Response intercept(@NonNull Chain chain) throws IOException {
-        Request request  = chain.request().newBuilder()
-                .addHeader("Authorization", "Bearer " + mAuthToken)
+
+        String keys = BuildConfig.CONSUMER_KEY + ":" + BuildConfig.CONSUMER_SECRET;
+
+        Request request = chain.request().newBuilder()
+                .addHeader("Authorization", "Basic " + Base64.encodeToString(keys.getBytes(), Base64.NO_WRAP))
                 .build();
         return chain.proceed(request);
     }
